feat: annotate elements
Optional elements and attributes are indicated in the annotations.
Default value is indicated in the annotations.
Min and max values are indicated in the annotations. They are checked by
the XML validator.
Equivalent elements are indicated in the annotations.
Bullet recommendations are provided in the annotations when they exist.
Values usage is sometimes indicated in the annotations.
My comments are provided in the annotations when something is
interpreted from the code.
Some more restrictive elements grouping (all instead of choice). This
better represent what is accepted by FSMP.